+++ src/dftbp/dftbplus/mainio.F90
@@ -10,6 +10,7 @@
 #! to a source code documentation tool.
 
 #:include 'common.fypp'
+#:include 'error.fypp'
 
 !> Various I/O routines for the main program.
 module dftbp_dftbplus_mainio
@@ -19,6 +20,7 @@ module dftbp_dftbplus_mainio
   use dftbp_common_environment, only : TEnvironment
   use dftbp_common_file, only : TFile, TFile_create, TFileOptions
   use dftbp_common_globalenv, only : stdOut, destructGlobalEnv, abortProgram
+  use dftbp_common_status, only : TStatus
   use dftbp_dftb_determinants, only : TDftbDeterminants
   use dftbp_dftb_dispersions, only : TDispersionIface
   use dftbp_dftb_elstatpot, only : TElStatPotentials
@@ -32,8 +34,8 @@ module dftbp_dftbplus_mainio
   use dftbp_extlibs_xmlf90, only : xmlf_t, xml_OpenFile, xml_ADDXMLDeclaration, xml_NewElement,&
       & xml_EndElement, xml_Close
   use dftbp_io_charmanip, only : i2c
-  use dftbp_io_commonformats, only : formatHessian, formatGeoOut, format1U, format2U, format1Ue,&
-      & format2Ue, format1U1e
+  use dftbp_io_commonformats, only : formatHessian, formatBorn, formatGeoOut, format1U, format2U,&
+      & format1Ue, format2Ue, format1U1e
   use dftbp_io_formatout, only : writeXYZFormat, writeGenFormat, writeSparse, writeSparseAsSquare
   use dftbp_io_hsdutils, only : writeChildValue
   use dftbp_io_message, only : error, warning
@@ -73,11 +75,11 @@ module dftbp_dftbplus_mainio
 #:endif
   public :: writeProjectedEigenvectors
   public :: initOutputFile, writeAutotestTag, writeResultsTag, writeDetailedXml, writeBandOut
-  public :: writeDerivBandOut, writeHessianOut
+  public :: writeDerivBandOut, writeHessianOut, writeBornChargesOut
   public :: openOutputFile
   public :: writeDetailedOut1, writeDetailedOut2, writeDetailedOut2Dets, writeDetailedOut3
-  public :: writeDetailedOut4, writeDetailedOut5, writeDetailedOut6
-  public :: writeDetailedOut7, writeDetailedOut8, writeDetailedOut9
+  public :: writeDetailedOut4, writeDetailedOut5, writeDetailedOut6, writeDetailedOut7
+  public :: writeDetailedOut8, writeDetailedOut9, writeDetailedOut10
   public :: writeMdOut1, writeMdOut2
   public :: writeCharges
   public :: writeEsp
@@ -2050,9 +2052,9 @@ contains
 
   !> Writes out machine readable data
   subroutine writeResultsTag(fileName, energy, derivs, chrgForces, nEl, Ef, eigen, filling,&
-      & electronicSolver, tStress, totalStress, pDynMatrix, tPeriodic, cellVol, tMulliken,&
-      & qOutput, q0, taggedWriter, cm5Cont, polarisability, dEidE, dqOut, neFermi, dEfdE,&
-      & coord0, dipoleMoment, multipole, eFieldScaling)
+      & electronicSolver, tStress, totalStress, pDynMatrix, pBornMatrix, tPeriodic, cellVol,&
+      & tMulliken, qOutput, q0, taggedWriter, cm5Cont, polarisability, dEidE, dqOut, neFermi,&
+      & dEfdE, coord0, dipoleMoment, multipole, eFieldScaling)
 
     !> Name of output file
     character(*), intent(in) :: fileName
@@ -2089,6 +2091,9 @@ contains
 
     !> Hessian (dynamical) matrix
     real(dp), pointer, intent(in) :: pDynMatrix(:,:)
+
+    !> Born charge matrix
+    real(dp), pointer, intent(in) :: pBornMatrix(:,:)
 
     !> Is the geometry periodic
     logical, intent(in) :: tPeriodic
@@ -2184,6 +2189,10 @@ contains
     end if
     if (associated(pDynMatrix)) then
       call taggedWriter%write(fd, tagLabels%HessianNum, pDynMatrix)
+    end if
+    if (associated(pBornMatrix)) then
+      call taggedWriter%write(fd, tagLabels%BorndDipNum,&
+          & eFieldScaling%scaledSoluteDipole(pBornMatrix))
     end if
     if (tPeriodic) then
       call taggedWriter%write(fd, tagLabels%volume, cellVol)
@@ -2483,8 +2492,8 @@ contains
   end subroutine writeDBand
 
 
-  !> Write the second derivative matrix
-  subroutine writeHessianOut(fileName, pDynMatrix, indMovedAtoms)
+  !> Write the energy second derivative matrix
+  subroutine writeHessianOut(fileName, pDynMatrix, indMovedAtoms, errStatus)
 
     !> File name
     character(*), intent(in) :: fileName
@@ -2495,15 +2504,17 @@ contains
     !> Indices of moved atoms
     integer, intent(in) :: indMovedAtoms(:)
 
+    !> Status of operation
+    type(TStatus), intent(out) :: errStatus
 
     integer :: ii, fd
     character(10) :: suffix1, suffix2
-    logical :: tPartialHessian = .false. 
+    logical :: tPartialHessian = .false.
 
     ! Sanity check in case some bug is introduced
     if (size(pDynMatrix, dim=2) /= 3*size(indMovedAtoms)) then
-      call error('Internal error: incorrect number of rows of dynamical Matrix')    
-    end if       
+      @:RAISE_ERROR(errStatus, -1, "Internal error: incorrect number of rows of dynamical Matrix")
+    end if
     ! It is a partial Hessian Calculation if DynMatrix is not squared
     if (size(pDynMatrix, dim=1) > size(pDynMatrix, dim=2)) then
       tPartialHessian = .true.
@@ -2511,10 +2522,10 @@ contains
 
     if (tPartialHessian) then
       write(suffix1,'(I10)') indMovedAtoms(1)
-      write(suffix2,'(I10)') indMovedAtoms(size(indMovedAtoms))     
-      open(newunit=fd, file=fileName//"."//trim(adjustl(suffix1))//"-"//trim(adjustl(suffix2)), &
-            & action="write", status="replace")
-    else 
+      write(suffix2,'(I10)') indMovedAtoms(size(indMovedAtoms))
+      open(newunit=fd, file=fileName//"."//trim(adjustl(suffix1))//"-"//trim(adjustl(suffix2)),&
+          & action="write", status="replace")
+    else
       open(newunit=fd, file=fileName, action="write", status="replace")
     end if
 
@@ -2525,13 +2536,70 @@ contains
     close(fd)
 
     if (tPartialHessian) then
-      write(stdOut, "(2A)") 'Hessian matrix written to ', &
-            & fileName//"."//trim(adjustl(suffix1))//"-"//trim(adjustl(suffix2))
+      write(stdOut, "(2A)") 'Hessian matrix written to ',&
+          & fileName//"."//trim(adjustl(suffix1))//"-"//trim(adjustl(suffix2))
     else
       write(stdOut, "(2A)") 'Hessian matrix written to ', fileName
     end if
 
   end subroutine writeHessianOut
+
+
+  !> Write the dipole derivative wrt.coordinates matrix/Born charges
+  subroutine writeBornChargesOut(fileName, pBornMatrix, indMovedAtoms, nAtInCentralRegion,&
+      & errStatus)
+
+    !> File name
+    character(*), intent(in) :: fileName
+
+    !> Born (dipole derivatives or force wrt electric field)
+    real(dp), intent(in) :: pBornMatrix(:,:)
+
+    !> Indices of moved atoms
+    integer, intent(in) :: indMovedAtoms(:)
+
+    !> Number of atoms in central region
+    integer, intent(in) :: nAtInCentralRegion
+
+    !> Status of operation
+    type(TStatus), intent(out) :: errStatus
+
+    integer :: ii, fd
+    character(10) :: suffix1, suffix2
+    logical :: tPartialMatrix = .false.
+
+    ! Sanity check in case some bug is introduced
+    if (any(shape(pBornMatrix) /= [3,3*size(indMovedAtoms)])) then
+      @:RAISE_ERROR(errStatus, -1, "Internal error: incorrectly shaped Born Matrix")
+    end if
+    ! It is a partial matrix Calculation if BornMatrix is not squared
+    if (size(pBornMatrix, dim=2) > 3*nAtInCentralRegion) then
+      tPartialMatrix = .true.
+    end if
+
+    if (tPartialMatrix) then
+      write(suffix1,'(I10)') indMovedAtoms(1)
+      write(suffix2,'(I10)') indMovedAtoms(size(indMovedAtoms))
+      open(newunit=fd, file=fileName//"."//trim(adjustl(suffix1))//"-"//trim(adjustl(suffix2)),&
+          & action="write", status="replace")
+    else
+      open(newunit=fd, file=fileName, action="write", status="replace")
+    end if
+
+    do ii = 1, size(pBornMatrix, dim=2)
+      write(fd, formatBorn) pBornMatrix(:, ii)
+    end do
+
+    close(fd)
+
+    if (tPartialMatrix) then
+      write(stdOut, "(2A)") 'Born charges matrix written to ',&
+          & fileName//"."//trim(adjustl(suffix1))//"-"//trim(adjustl(suffix2))
+    else
+      write(stdOut, "(2A)") 'Born charges matrix written to ', fileName
+    end if
+
+  end subroutine writeBornChargesOut
 
 
   !> Opens an output file or uses the its current unit number, if the file is already open.
@@ -3669,9 +3737,28 @@ contains
 
   end subroutine writeDetailedOut7
 
+  !> Eighth group of data for detailed.out (Born effective charges)
+  subroutine writeDetailedOut8(fd, born)
 
-  !> Eighth group of data for detailed.out (density of states at Fermi energy)
-  subroutine writeDetailedOut8(fd, neFermi)
+    !> File ID
+    integer, intent(in) :: fd
+
+    !> Born charges
+    real(dp), intent(in) :: born(:,:)
+
+    integer :: ii
+
+    write(fd,*)'Born charges/dipole derivatives wrt. atom positions (e)'
+    do ii = 1, size(born,dim=2), 3
+      write(fd,"(A,1X,I0)")'Atom',ii/3+1
+      write(fd,"(3F12.6)")born(:,ii:ii+2)
+    end do
+
+  end subroutine writeDetailedOut8
+
+
+  !> Nineth group of data for detailed.out (density of states at Fermi energy)
+  subroutine writeDetailedOut9(fd, neFermi)
 
     !> File ID
     integer, intent(in) :: fd
@@ -3688,11 +3775,11 @@ contains
       end if
     end if
 
-  end subroutine writeDetailedOut8
+  end subroutine writeDetailedOut9
 
 
-  !> Nineth group of data for detailed.out (derivatives with respect to an external electric field)
-  subroutine writeDetailedOut9(fd, orb, polarisability, dqOut, dEfdE)
+  !> Tenth group of data for detailed.out (derivatives with respect to an external electric field)
+  subroutine writeDetailedOut10(fd, orb, polarisability, dqOut, dEfdE)
 
     !> File ID
     integer, intent(in) :: fd
@@ -3767,7 +3854,7 @@ contains
       write(fd,*)
     end if
 
-  end subroutine writeDetailedOut9
+  end subroutine writeDetailedOut10
 
 
   !> First group of output data during molecular dynamics